pyproject.toml
[project]
name = "courseconcierge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "langchain-anthropic",
    "playwright",
]
